train_models/train_plms/requirements.txt
absl-py==1.1.0
accelerate==0.26.1
aiohttp==3.9.1
aiosignal==1.3.1
apex==0.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
boto3==1.24.7
botocore==1.27.7
cachetools==5.2.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
datasets==2.16.1
decord==0.6.0
dill==0.3.7
einops==0.4.1
evaluate==0.4.1
expecttest==0.1.3
filelock==3.7.1
frozenlist==1.4.1
fsspec==2023.10.0
ftfy==6.1.1
future==0.18.2
google-auth==2.7.0
google-auth-oauthlib==0.4.6
grpcio==1.46.3
huggingface-hub==0.20.2
idna==3.3
importlib-metadata==4.11.4
jmespath==1.0.0
joblib==1.3.2
Markdown==3.3.7
multidict==6.0.4
multiprocess==0.70.15
numpy==1.22.4
oauthlib==3.2.0
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
packaging==21.3
pandas==1.4.2
Pillow==9.1.1
protobuf==3.19.4
psutil==5.9.1
pyarrow==8.0.0
pyarrow-hotfix==0.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.9.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
regex==2022.6.2
requests==2.28.0
requests-oauthlib==1.3.1
responses==0.18.0
rsa==4.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.6.0
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.10.1
sentencepiece==0.1.99
six==1.16.0
sox==1.4.1
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
termcolor==1.1.0
threadpoolctl==3.2.0
timm==0.5.4
tokenizers==0.15.0
torch @ file:///root/pytorch/dist/torch-1.11.0-cp38-cp38-linux_x86_64.whl
torchvision @ file:///root/vision/dist/torchvision-0.12.0-cp38-cp38-linux_x86_64.whl
tqdm==4.64.0
transformers @ file:///workspace/dissertation/transformers
types-dataclasses==0.6.5
typing_extensions==4.2.0
urllib3==1.26.9
wcwidth==0.2.5
Werkzeug==2.1.2
xxhash==3.4.1
yarl==1.9.4
zipp==3.8.0
